reorganise bitmatrix code for maintainability
