--- python/triton_kernels/triton_kernels/__init__.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class Bitmatrix:
-    data: "torch.Tensor"  # noqa: F821
-    shape: tuple[int]
